LoopStream クラスを別ファイルに移動
